+++ src/main/java/ru/skypro/homework/dto/CommentDto.java
@@ -9,7 +9,7 @@ public class CommentDto {
     private String authorImage;     // ссылка на аватар автора комментария
     private String authorFirstName; // имя создателя комментария
     private Long createdAt;         //дата и время создания комментария в миллисекундах с 00:00:00 01.01.1970
-    private Integer pk;                // id комментария
+    private Long pk;                // id комментария
     private String text;            // текст комментария
 
 }
